requirements.txt
Django==2.2
PyJWT==1.7.1
PyMySQL==0.9.3
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
django-forms-bootstrap==3.1.0
django-tinymce4-lite==1.7.5
idna==2.8
jsmin==2.2.2
oauthlib==3.0.1
psycopg2==2.8.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
